hide-owner-email - move owner-contact-details to bark/models
