+++ src/app/vet/_lib/models/scheduler-state.ts
@@ -1,4 +1,5 @@
-import { AvailableDog, OwnerContactDetails } from "@/lib/vet/vet-models";
+import { AvailableDog } from "@/lib/vet/vet-models";
+import { OwnerContactDetails } from "@/lib/bark/models/owner-contact-details";
 import { SchedulerOutcome } from "./scheduler-outcome";
 
 export type SchedulerState = {
