+++ src/lib/vet/vet-models.ts
@@ -11,15 +11,3 @@ export type AvailableDog = {
   dogEverReceivedTransfusion: YesNoUnknown;
   dogEverPregnant: YesNoUnknown;
 };
-
-export type OwnerContactDetails = {
-  dogId: string;
-  userName: string;
-  userEmail: string;
-  userPhoneNumber: string;
-
-  /**
-   * The last time the vet contacted the user.
-   */
-  vetUserLastContactedTime: Date | null;
-};
